broken checkpoint commit
I needed to transfer to working on my laptop, windows 11 is being a pain. broken commit, current push hits segment fault when making the graph causing the program to crash. As plotting the lines based on trade data already existing apparently is using too much memory? it works somewhat. i need to take a break on this for the moment though. left out all json and csv from push